feat: implement onboarding animation and language selection flow
