+++ features/login.feature
@@ -4,12 +4,6 @@ Feature: Login Feature (existing)
   So that I can access the main features of the app
   
 
-  @regression @onboarding @login
-  Scenario: Display onboarding animation and start using the app
-    Given the app is launched for the first time
-    When I select my language and click sure button
-    Then I can start using the app
-
   @regression @select_language_and_contact_cs @login
   Scenario: Select language and click contact cs
     Given the app is launched
