+++ pages/android/login_page.py
@@ -9,41 +9,51 @@ from pages.shared_components.common_action import CommonActions
 class LoginPage(CommonActions):
   def __init__(self, driver):
       super().__init__(driver)
-  
+
   def select_language(self):
+      self.wait_for_element_visible(*LoginLocators.LANGUAGE_SETTING_BUTTON)
       self.click_element(*LoginLocators.LANGUAGE_SETTING_BUTTON)
+      self.wait_for_element_visible(*LoginLocators.CHINESE_LANGUAGE)
       self.click_element(*LoginLocators.CHINESE_LANGUAGE)
+      self.wait_for_element_visible(*LoginLocators.LANGUAGE_SAVE_BUTTON)
       self.click_element(*LoginLocators.LANGUAGE_SAVE_BUTTON)
 
   def continue_to_login_page(self):
+      self.wait_for_element_visible(*LoginLocators.LANGUAGE_SETTING_BUTTON)
       language_setting_btn = self.find_element(*LoginLocators.LANGUAGE_SETTING_BUTTON)
       contact_cs_btn = self.find_element(*LoginLocators.CONTACT_CS_BUTTON)
-      
+
       assert language_setting_btn.is_displayed(), "Language setting button not found"
       assert contact_cs_btn.is_displayed(), "Contact customer service button not found"
-      assert self.find_element(*LoginLocators.LOGIN_BUTTON) is not None, "Start using the app button not found"
-      
+      assert self.wait_for_element_visible(*LoginLocators.LOGIN_BUTTON) is not None, "Start using the app button not found"
+
   def click_contact_cs_button(self):
+      self.wait_for_element_visible(*LoginLocators.CONTACT_CS_BUTTON)
       self.click_element(*LoginLocators.CONTACT_CS_BUTTON)
+      self.wait_for_element_visible(*LoginLocators.CONTACT_CS_BACK_BUTTON)
       self.click_element(*LoginLocators.CONTACT_CS_BACK_BUTTON)
 
-      
+
   def click_terms_and_conditions_button(self):
+      self.wait_for_element_visible(*LoginLocators.TERMS_AND_CONDITIONS_BUTTON)
       self.click_element(*LoginLocators.TERMS_AND_CONDITIONS_BUTTON)
-  
+
   def click_tc_back_button(self):
+      self.wait_for_element_visible(*LoginLocators.TC_BACK_BUTTON)
       self.click_element(*LoginLocators.TC_BACK_BUTTON)
 
   def click_privacy_button(self):
+      self.wait_for_element_visible(*LoginLocators.PRIVACY_BUTTON)
       self.click_element(*LoginLocators.PRIVACY_BUTTON)
 
   def click_privacy_back_button(self):
+      self.wait_for_element_visible(*LoginLocators.PRIVACY_BACK_BUTTON)
       self.click_element(*LoginLocators.PRIVACY_BACK_BUTTON)
 
   def click_login_button(self):
       self.wait_for_element_clickable(*LoginLocators.LOGIN_BUTTON)
       self.click_element(*LoginLocators.LOGIN_BUTTON)
-  
+
   def enter_email(self, email):
       try:
           email_input = self.find_element(*LoginLocators.EMAIL_INPUT)
@@ -54,7 +64,7 @@ class LoginPage(CommonActions):
             return
       except NoSuchElementException:
         pass
-      
+
       try:
           self.send_keys_to_element(*LoginLocators.EMAIL_INPUT, email)
           self.click_element(*LoginLocators.EMAIL_NEXT_BUTTON)
@@ -86,10 +96,10 @@ class LoginPage(CommonActions):
           return None
       except NoSuchElementException:
           return False
-  
+
   def login_with_unregistered_email(self, email):
        self.click_login_button()
-    
+
        try:
             email_input = self.find_element(*LoginLocators.EMAIL_INPUT)
             current_value = email_input.get_attribute("text")
@@ -99,12 +109,12 @@ class LoginPage(CommonActions):
             email_input.send_keys(email)
             self.click_element(*LoginLocators.EMAIL_NEXT_BUTTON)
 
-            
+
        except NoSuchElementException as e:
             print(f"無法找到元素: {str(e)}")
-            raise  
-        
-      
+            raise
+
+
   def error_unregistered_message(self):
       error_element = self.find_element(*LoginLocators.ERROR_UNREGISTERED_WINDOW_TITLE)
       assert error_element.is_displayed(), "Error unregistered window title not found"
@@ -112,32 +122,32 @@ class LoginPage(CommonActions):
       self.click_element(*LoginLocators.ERROR_UNREGISTERED_WINDOW_BUTTON)
       return error_text
 
-  
-  
+
+
   def get_email_error_message(self):
-       try: 
+       try:
             error_element = self.find_element(*LoginLocators.ERROR_MESSAGE)
             return error_element.text.strip()
-        
+
        except NoSuchElementException:
             return None
-        
-       
-        
+
+
+
   def error_ver_window(self):
         try:
             error_element = self.find_element(*LoginLocators.ERROR_WINDOW_TEXT)
             error_text = error_element.text.strip()
             self.click_element(*LoginLocators.ERROR_RETRY_BUTTON)
             return error_text
-            
+
         except NoSuchElementException:
             return None
-       
+
   def click_login_cancel_button(self):
        self.click_element(*LoginLocators.LOGIN_CANCEL_BUTTON)
-       
- 
+
+
   def login(self, email, ver_code=None):
         '''run the valid login process'''
         self.click_login_button()
@@ -149,23 +159,23 @@ class LoginPage(CommonActions):
             current_value = email_input.get_attribute("text")
             if current_value:
                 email_input.clear()
-            
+
             # enter email
             self.send_keys_to_element(*LoginLocators.EMAIL_INPUT, email)
-            
+
             # click next button
             self.wait_for_element_clickable(*LoginLocators.EMAIL_NEXT_BUTTON)
             self.click_element(*LoginLocators.EMAIL_NEXT_BUTTON)
-            
+
             try:
                 # verification code
                 self.enter_ver_code(ver_code)
-                
+
             except:
                 self.click_finish_button()
-            
+
             return True
-            
+
         except NoSuchElementException as e:
             print(f"無法找到元素: {str(e)}")
             raise
@@ -183,17 +193,17 @@ class LoginPage(CommonActions):
             current_value = email_input.get_attribute("text")
             if current_value:
                 email_input.clear()
-            
+
             # enter email
             email_input.send_keys(email)
-            
+
             # click next button
             self.click_element(*LoginLocators.EMAIL_NEXT_BUTTON)
 
         except Exception as e:
             print(f"發生未預期的錯誤: {str(e)}")
             raise
-        
+
   def login_with_another_valid_email(self, email, ver_code=None):
         '''run the valid login process'''
 
@@ -203,20 +213,20 @@ class LoginPage(CommonActions):
             current_value = email_input.get_attribute("text")
             if current_value:
                 email_input.clear()
-            
+
             # enter email
             email_input.send_keys(email)
-            
+
             # click next button
             self.click_element(*LoginLocators.EMAIL_NEXT_BUTTON)
 
         except Exception as e:
             print(f"發生未預期的錯誤: {str(e)}")
             raise
-        
+
   def click_modify_email_button(self):
         self.click_element(*LoginLocators.MODIFY_EMAIL_BUTTON)
-        
+
   def is_verification_code_page(self):
       verification_code_page_title = self.find_element(*LoginLocators.IS_VERIFICATION_CODE_PAGE_TITLE)
       return verification_code_page_title.is_displayed()
@@ -224,7 +234,7 @@ class LoginPage(CommonActions):
   def login_with_invalid_email(self, email):
         """run the invalid login process"""
         self.click_login_button()
-    
+
         try:
             email_input = self.find_element(*LoginLocators.EMAIL_INPUT)
             current_value = email_input.get_attribute("text")
@@ -233,15 +243,15 @@ class LoginPage(CommonActions):
 
             email_input.send_keys(email)
 
-            
+
         except NoSuchElementException as e:
             print(f"無法找到元素: {str(e)}")
-            raise  
-        
+            raise
+
   def login_with_invalid_ver_code(self, ver_code=None):
         """run the invalid login process"""
         self.enter_ver_code(ver_code)
-        
+
 
 ########### Just for logout ###########
   def click_logout_button(self):
